pyproject.toml
[project]
name = "eurosat-mlops-pipeline"
version = "0.1.0"
description = "Minimal, traceable EuroSAT train/eval/predict pipeline"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "pillow>=10.0",
  "pyyaml>=6.0",
  "torch",
  "torchvision",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.ruff]
line-length = 100

target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]

[tool.setuptools]
packages = ["src"]  

[tool.coverage.run]
source = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
